// src/types/store/mutation.types.ts
export enum AuthMutationTypes {
  SET_NEW_TOKEN = "SET_NEW_TOKEN",
  SET_USER = "SET_USER",
  SET_AUTH = "SET_AUTH",
  PURGE_AUTH = "PURGE_AUTH",
  SET_GLOBAL_LOADING = "SET_GLOBAL_LOADING",
  SET_MOUNT = "SET_MOUNT",
}

export enum PropertyMutationTypes {
  SET_PROPERTY_CLASSIFICATIONS = "SET_PROPERTY_CLASSIFICATIONS",
  SET_PROPERTY_TYPES = "SET_PROPERTY_TYPES",
  SET_PROPERTY_FUNDS = "SET_PROPERTY_FUNDS",
  SET_PROPERTY_SELLINGBANKS = "SET_PROPERTY_SELLINGBANKS",
  SET_COUNTRIES = "SET_COUNTRIES",
  SET_STATES = "SET_STATES",
  SET_CITIES = "SET_CITIES",
  SET_STATUS = "SET_STATUS",
  SET_AMENITIES = "SET_AMENITIES",
}

export enum InvestorMutationTypes {
  SET_INVESTOR_PROFILE = "SET_INVESTOR_PROFILE",
  SET_INVESTOR_FUND = "SET_INVESTOR_FUND",
  SET_PROFILE_ID = "SET_PROFILE_ID",
  SET_CLIENT_LOADING = "SET_CLIENT_LOADING",
}